fix: use type long long for n=100000000

+++ task1.c
@@ -2,9 +2,9 @@
 
 int main() {
     double approx = 0;
-    double n = 100000000;
+    long long n = 100000000;
 
-    for (int i = 0; i <= n; i++) {
+    for (long long i = 0; i <= n; i++) {
         double numerator = 1;
         if (i%2 == 1) {
             numerator *= -1;
